refactor(Friend): API 분리, 친구 카드 스타일 수정
